backend/requirements.txt
Django==6.0.1
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-decouple==3.8
djangorestframework-simplejwt==5.5.1
paho-mqtt==2.1.0
Pillow>=10.0.0
